+++ game/ground/IGroundHelper.cs
@@ -22,11 +22,11 @@ namespace AbrahmanAdventure.physics
         internal static IGround GetHighestVisibleIGroundBelowSprite(AbstractSprite sprite, Level level, HashSet<AbstractSprite> visibleSpriteList)
         {
             IGround highestIGroundBelowSprite = null;
-            float highestHeight = -1;
+            double highestHeight = -1;
 
             foreach (Ground ground in level)
             {
-                float currentHeight = ground[sprite.XPosition];
+                double currentHeight = ground[sprite.XPosition];
 
                 if (sprite.YPosition <= currentHeight)
                 {
@@ -53,7 +53,7 @@ namespace AbrahmanAdventure.physics
 
                     if (isHorizontalCollision)
                     {
-                        float currentHeight = otherSprite.TopBound;
+                        double currentHeight = otherSprite.TopBound;
                         if (sprite.YPosition <= currentHeight)
                         {
                             if (otherSprite.TopBound <= currentHeight)
@@ -79,9 +79,9 @@ namespace AbrahmanAdventure.physics
         /// <param name="level">level</param>
         /// <param name="xPosition">X Position</param>
         /// <returns>Whether ground is visible at X Position</returns>
-        internal static bool IsGroundVisible(Ground ground, Level level, float xPosition)
+        internal static bool IsGroundVisible(Ground ground, Level level, double xPosition)
         {
-            float yPosition = ground[xPosition];
+            double yPosition = ground[xPosition];
 
             for (int groundId = level.Count - 1; groundId >= 0; groundId--)
             {
@@ -105,11 +105,11 @@ namespace AbrahmanAdventure.physics
         internal static Ground GetLowestVisibleGround(AbstractSprite sprite, Level level)
         {
             Ground lowestGround = null;
-            float lowestHeight = float.NegativeInfinity;
+            double lowestHeight = double.NegativeInfinity;
 
             foreach (Ground ground in level)
             {
-                float currentHeight = ground[sprite.XPosition];
+                double currentHeight = ground[sprite.XPosition];
 
                 if (currentHeight > lowestHeight)
                 {
@@ -130,7 +130,7 @@ namespace AbrahmanAdventure.physics
         /// <param name="level">level</param>
         /// <param name="x">x position</param>
         /// <returns></returns>
-        internal static bool IsTransparentAt(Ground ground, Level level, float x)
+        internal static bool IsTransparentAt(Ground ground, Level level, double x)
         {
         	return ground.IsTransparent && IGroundHelper.IsHigherThanOtherGrounds(ground, level, x);
         }
@@ -141,9 +141,9 @@ namespace AbrahmanAdventure.physics
         /// <param name="level">level</param>
         /// <param name="xInput">x value</param>
         /// <returns>Whether ground is higher than other grounds</returns>
-        internal static bool IsHigherThanOtherGrounds(Ground ground, Level level, float xInput)
+        internal static bool IsHigherThanOtherGrounds(Ground ground, Level level, double xInput)
         {
-            float yOutput = ground[xInput];
+            double yOutput = ground[xInput];
             foreach (Ground otherGround in level)
                 if (otherGround != ground)
                     if (otherGround[xInput] < yOutput)
@@ -159,7 +159,7 @@ namespace AbrahmanAdventure.physics
         /// <param name="waveInputX">x input of wave</param>
         /// <param name="yOutput">wave y output</param>
         /// <returns>Whether ground is higher than other grounds that are in front of it</returns>
-        internal static bool IsHigherThanOtherGroundsInFront(Ground ground, Level level, float waveInputX, float yOutput)
+        internal static bool IsHigherThanOtherGroundsInFront(Ground ground, Level level, double waveInputX, double yOutput)
         {
             bool isCrossedSourceGround = false;
             foreach (Ground otherGround in level)
@@ -187,7 +187,7 @@ namespace AbrahmanAdventure.physics
         internal static IGround GetFrontmostGroundHavingAccessibleWalkingHeightForSprite(AbstractSprite sprite, IGround iGround, Level level, HashSet<AbstractSprite> visibleSpriteList)
         {
             IGround frontMostGroundHavingAccessibleWalkingHeightForSprite = null;
-            float groundHeight = iGround[sprite.XPosition];
+            double groundHeight = iGround[sprite.XPosition];
 
             for (int groundId = level.Count - 1; groundId >= 0; groundId--)
             {
@@ -196,7 +196,7 @@ namespace AbrahmanAdventure.physics
                 if (currentGround == iGround)
                     break;
 
-                float currentGroundHeight = currentGround[sprite.XPosition];
+                double currentGroundHeight = currentGround[sprite.XPosition];
 
                 if (currentGroundHeight < groundHeight && groundHeight - currentGroundHeight <= sprite.MaximumWalkingHeight)
                 {
@@ -211,7 +211,7 @@ namespace AbrahmanAdventure.physics
                 {
                     if ((sprite.RightBound > otherSprite.LeftBound && sprite.LeftBound < otherSprite.LeftBound) || (sprite.LeftBound < otherSprite.RightBound && sprite.LeftBound > otherSprite.LeftBound))
                     {
-                        float currentHeight = otherSprite.TopBound;
+                        double currentHeight = otherSprite.TopBound;
                         if (sprite.YPosition <= currentHeight)
                         {
                             if (Math.Abs(groundHeight - currentHeight) <= sprite.MaximumWalkingHeight)
@@ -237,13 +237,13 @@ namespace AbrahmanAdventure.physics
         /// <returns>highest ground having accessible walking height for sprite</returns>
         internal static IGround GetHighestGroundHavingAccessibleWalkingHeightForSprite(AbstractSprite sprite, IGround iGround, Level level, HashSet<AbstractSprite> visibleSpriteList)
         {
-        	float groundHeight = iGround[sprite.XPosition];
-            float highestHeight = float.PositiveInfinity;
+        	double groundHeight = iGround[sprite.XPosition];
+            double highestHeight = double.PositiveInfinity;
             IGround highestGround = null;
 
             foreach (Ground currentGround in level)
             {
-                float currentGroundHeight = currentGround[sprite.XPosition];
+                double currentGroundHeight = currentGround[sprite.XPosition];
 
                 if (currentGroundHeight < highestHeight && Math.Abs(groundHeight - currentGroundHeight) <= sprite.MaximumWalkingHeight)
                 {
@@ -258,7 +258,7 @@ namespace AbrahmanAdventure.physics
                 {
                     if ((sprite.RightBound > otherSprite.LeftBound && sprite.LeftBound < otherSprite.LeftBound) || (sprite.LeftBound < otherSprite.RightBound && sprite.LeftBound > otherSprite.LeftBound))
                     {
-                        float currentHeight = otherSprite.TopBound;
+                        double currentHeight = otherSprite.TopBound;
                         if (sprite.YPosition <= currentHeight)
                         {
                             if (otherSprite.TopBound <= currentHeight)
